requirements.txt
Flask==2.3.3
langchain-nvidia-ai-endpoints==0.0.1
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
Werkzeug==2.3.7